# lib/api-zod/src/generated/types/chatMessageInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SentinelGrid API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessageInput {
  message: string;
  /** @nullable */
  role?: string | null;
}