# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SentinelGrid API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CountByLabel } from './countByLabel';

export interface DashboardSummary {
  activeIncidents: number;
  criticalAlerts: number;
  riskScore: number;
  mttd: number;
  mttr: number;
  entitiesMonitored: number;
  containedThreats: number;
  alertsByStage: CountByLabel[];
  alertsBySeverity: CountByLabel[];
}