# lib/api-zod/src/generated/types/playbookStep.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SentinelGrid API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlaybookStepStatus } from './playbookStepStatus';

export interface PlaybookStep {
  step: number;
  action: string;
  requiresApproval: boolean;
  status: PlaybookStepStatus;
}